refactor(types): move OScriptLibraryIndexedEvent to events package
Событие переезжает в новый пакет types.oscript.events рядом с остальными
Spring ApplicationEvent проекта. Это приводит размещение события к общей
конвенции (все события — в пакетах events) и снимает единственное исключение
для двунаправленного архитектурного правила про события.
